+++ forge/src/main/java/smartin/miapi/forge/compat/quark/QuarkClientCompat.java
@@ -1,4 +1,4 @@
-package smartin.miapi.forge.compat;
+package smartin.miapi.forge.compat.quark;
 
 import com.redpxnda.nucleus.util.Color;
 import dev.architectury.event.EventResult;
@@ -11,7 +11,7 @@ import smartin.miapi.modules.properties.GlintProperty;
 
 import java.util.concurrent.atomic.AtomicReference;
 
-public class QuarkCompat {
+public class QuarkClientCompat {
 
     public static void setup() {
         GlintProperty.GLINT_RESOLVE.register(new GlintProperty.GlintGetter() {