+++ forge/src/main/java/smartin/miapi/forge/compat/quark/QuarkCompat.java
@@ -1,0 +1,12 @@
+package smartin.miapi.forge.compat.quark;
+
+import smartin.miapi.Environment;
+
+public class QuarkCompat {
+
+    public static void setup() {
+        if (Environment.isClient()) {
+            smartin.miapi.forge.compat.quark.QuarkClientCompat.setup();
+        }
+    }
+}